Fix file headers and footer for ELPA compatibility
The problems fixed by this commit are the reason `evil-leader` barely scrapes into Melpa with a "No description available" comment. :-)

Related: a good test of whether an .el file is well-formatted for publishing to an ELPA archive is to run `M-: (package-buffer-info)` -- if no error is raised, you're pretty much good to go. :-)